+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## unreleased
+## 6.5.0
 * Add plan create/update/find API endpoint
 * Add `TransactionReview` webhook notification